64 bit for expression, fixes issue #53
(cherry picked from commit 080dc529e4d3ee621ebac96691e1bbee95125058)

+++ src/device.c
@@ -27,6 +27,7 @@
 
 
 #include <stdlib.h>
+#include <inttypes.h>
 #include <string.h>
 
 #include "misc.h"
@@ -189,7 +190,7 @@ struct device *get_device(struct prog_info *pi, char *name)
 int
 predef_dev(struct prog_info *pi)
 {
-	int i;
+	int64_t i;
 	char temp[MAX_DEV_NAME+1];
 	def_dev(pi);
 	for (i=0; (!i)||(device_list[i].name); i++) {
@@ -206,13 +207,13 @@ predef_dev(struct prog_info *pi)
 			if (def_const(pi, temp, i)==False)
 				return (False);
 		} else { /* Pass 2 */
-			int j;
+			int64_t j;
 			if (get_constant(pi, temp, &j)==False) {  /* Defined in Pass 1 and now missing ? */
 				fprintf(stderr,"Constant %s is missing in pass 2\n",temp);
 				return (False);
 			}
 			if (i != j) {
-				fprintf(stderr,"Constant %s changed value from %d in pass1 to %d in pass 2\n",temp,j,i);
+				fprintf(stderr,"Constant %s changed value from %"PRId64" in pass1 to %"PRId64" in pass 2\n",temp,j,i);
 				return (False);
 			}
 			/* OK. definition is unchanged */
